Minor improvements & fixes

+++ builds/builds.go
@@ -69,47 +69,52 @@ func (build Build) CheckBuild(currentVersion *version.Version) (version.Version,
 	}
 
 	if remoteUrl != "" {
-		log.Printf("Downloading %s for new build of %s at version %s\n", remoteUrl, build.Name, newVersion)
-		localFile, err := files.DownloadFile(remoteUrl, os.TempDir())
-		if err != nil {
-			return *currentVersion, err
-		}
-		log.Printf("New build for %s was downloaded, installing from local file %s\n", build.Name, localFile)
-		if strings.HasSuffix(localFile, ".7z") || strings.HasSuffix(localFile, ".zip") {
-			entries, err := sevenzip.ReadEntries(localFile)
-			var commonFolderName string
+
+		if build.Location.Folder == "" {
+			log.Printf("%s has not location set, skipping installation", build.Name)
+		} else {
+
+			log.Printf("Downloading %s for new build of %s at version %s\n", remoteUrl, build.Name, newVersion)
+			localFile, err := files.DownloadFile(remoteUrl, os.TempDir())
 			if err != nil {
 				return *currentVersion, err
 			}
-			if strings.HasPrefix(entries[0].Attr, "D") {
-				commonFolderName = entries[0].Name
-				for _, entry := range entries[1:] {
-					if !strings.HasPrefix(entry.Name, commonFolderName) {
-						commonFolderName = ""
-						break
+			log.Printf("New build for %s was downloaded, installing from local file %s\n", build.Name, localFile)
+			if strings.HasSuffix(localFile, ".7z") || strings.HasSuffix(localFile, ".zip") {
+				entries, err := sevenzip.ReadEntries(localFile)
+				var commonFolderName string
+				if err != nil {
+					return *currentVersion, err
+				}
+				if strings.HasPrefix(entries[0].Attr, "D") {
+					commonFolderName = entries[0].Name
+					for _, entry := range entries[1:] {
+						if !strings.HasPrefix(entry.Name, commonFolderName) {
+							commonFolderName = ""
+							break
+						}
 					}
 				}
-			}
-			if commonFolderName != "" && build.Location.SuppressParentFolder {
-				unzipErr := sevenzip.Uncompress(localFile, os.TempDir())
-				if unzipErr == nil {
-					sourceFolder := path.Join(os.TempDir(), commonFolderName)
-					err = files.MoveFolder(sourceFolder, build.Location.Folder)
+				if commonFolderName != "" && build.Location.SuppressParentFolder {
+					unzipErr := sevenzip.Uncompress(localFile, os.TempDir())
+					if unzipErr == nil {
+						sourceFolder := path.Join(os.TempDir(), commonFolderName)
+						err = files.MoveFolder(sourceFolder, build.Location.Folder)
+					}
+				} else {
+					err = sevenzip.Uncompress(localFile, build.Location.Folder)
 				}
-			} else {
-				err = sevenzip.Uncompress(localFile, build.Location.Folder)
+				if err == nil {
+					os.Remove(localFile)
+				}
 			}
-			if err == nil {
-				os.Remove(localFile)
-			}
-		}
-
-		if build.PostInstallCmd != "" {
-			err = build.executePostBuildCommand(localFile)
-			if err != nil {
-				return *currentVersion, err
-			} else {
-				os.Remove(localFile)
+			if build.PostInstallCmd != "" {
+				err = build.executePostBuildCommand(localFile)
+				if err != nil {
+					return *currentVersion, err
+				} else {
+					os.Remove(localFile)
+				}
 			}
 		}
